+++ source/src/EnhancedCustomBattleConfig.cs
@@ -121,7 +121,7 @@ namespace EnhancedBattleTest
                     new ClassInfo { classStringId = "mp_heavy_ranged_battania", selectedFirstPerk = 0, selectedSecondPerk = 0, troopCount = 20 },
                     new ClassInfo { classStringId = "mp_heavy_infantry_battania", selectedFirstPerk = 0, selectedSecondPerk = 0, troopCount = 20 },
                 },
-                disableDying = false,
+                disableDeath = false,
                 changeCombatAI = false,
                 combatAI = 100,
             };