+++ source/src/EnhancedSiegeBattleViews.cs
@@ -47,7 +47,6 @@ namespace EnhancedBattleTest
                 ViewCreator.CreateMissionSingleplayerEscapeMenu(),
                 ViewCreator.CreateMissionOrderUIHandler(mission),
                 ViewCreator.CreateOrderTroopPlacerView(mission),
-                ViewCreator.CreateSingleplayerMissionKillNotificationUIHandler(),
                 new MissionItemContourControllerView(),
                 new MissionAgentContourControllerView(),
                 ViewCreator.CreateMissionBoundaryCrossingView(),
@@ -64,6 +63,12 @@ namespace EnhancedBattleTest
             {
                 missionViewList.Add(ViewCreator.CreateMissionAgentLabelUIHandler(mission));
             }
+
+            if (!config.noKillNotification)
+            {
+                missionViewList.Add(ViewCreator.CreateSingleplayerMissionKillNotificationUIHandler());
+            }
+
             return missionViewList.ToArray();
         }
     }